requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiosignal=1.3.1=pypi_0
aom=3.8.2=h63175ca_0
attrs=23.2.0=pypi_0
blas=2.112=mkl
blas-devel=3.9.0=12_win64_mkl
blinker=1.8.1=pypi_0
brotli=1.1.0=hcfcfb64_1
brotli-bin=1.1.0=hcfcfb64_1
brotli-python=1.1.0=py38hd3f51b4_1
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2024.3.11=haa95532_0
cairo=1.18.0=h1fef639_0
certifi=2024.2.2=py38haa95532_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.1.1=py38hb1fd069_1
cuda-cccl=12.4.127=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.4.127=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=hcfcfb64_0
double-conversion=3.3.0=h63175ca_0
expat=2.6.2=h63175ca_0
ffmpeg=6.1.1=gpl_h66c0b5b_108
filelock=3.14.0=pyhd8ed1ab_0
flask=3.0.3=pypi_0
flask-cors=4.0.0=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.51.0=py38h91455d4_0
freeglut=3.2.2=h63175ca_2
freetype=2.12.1=hdaf720e_2
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
gettext=0.22.5=h5728263_2
gettext-tools=0.22.5=h7d00a51_2
graphite2=1.3.13=h63175ca_1003
harfbuzz=8.4.0=h7ab893a_0
hdf5=1.14.3=nompi_h73e8ff5_100
icu=73.2=h63175ca_0
idna=3.7=pyhd8ed1ab_0
imath=3.1.11=h12be248_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pyhd8ed1ab_0
importlib_resources=6.4.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=pypi_0
itsdangerous=2.2.0=pypi_0
jasper=4.2.4=hcb1a123_0
jinja2=3.1.3=pyhd8ed1ab_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
khronos-opencl-icd-loader=2023.04.17=h64bf75a_0
kiwisolver=1.4.5=py38hb1fd069_1
krb5=1.21.2=heb0366b_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libabseil=20240116.2=cxx17_h63175ca_0
libaec=1.1.3=h63175ca_0
libasprintf=0.22.5=h5728263_2
libasprintf-devel=0.22.5=h5728263_2
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.1.0=hcfcfb64_1
libbrotlidec=1.1.0=hcfcfb64_1
libbrotlienc=1.1.0=hcfcfb64_1
libcblas=3.9.0=12_win64_mkl
libclang13=18.1.3=default_hf64faad_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcurl=8.7.1=hd5e4a3a_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.20=hcfcfb64_0
libexpat=2.6.2=h63175ca_0
libffi=3.4.4=hd77b12b_0
libgettextpo=0.22.5=h5728263_2
libgettextpo-devel=0.22.5=h5728263_2
libglib=2.80.0=h39d0aa6_6
libhwloc=2.10.0=default_h2fffb23_1000
libiconv=1.17=hcfcfb64_2
libintl=0.22.5=h5728263_2
libintl-devel=0.22.5=h5728263_2
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=12_win64_mkl
liblapacke=3.9.0=12_win64_mkl
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libopencv=4.9.0=qt6_py38he8dd393_612
libopenvino=2024.0.0=hc2557fa_4
libopenvino-auto-batch-plugin=2024.0.0=h002f227_4
libopenvino-auto-plugin=2024.0.0=h002f227_4
libopenvino-hetero-plugin=2024.0.0=h7e3b17c_4
libopenvino-intel-cpu-plugin=2024.0.0=hc2557fa_4
libopenvino-intel-gpu-plugin=2024.0.0=hc2557fa_4
libopenvino-ir-frontend=2024.0.0=h7e3b17c_4
libopenvino-onnx-frontend=2024.0.0=h55b4db4_4
libopenvino-paddle-frontend=2024.0.0=h55b4db4_4
libopenvino-pytorch-frontend=2024.0.0=h63175ca_4
libopenvino-tensorflow-frontend=2024.0.0=ha362bc9_4
libopenvino-tensorflow-lite-frontend=2024.0.0=h63175ca_4
libopus=1.3.1=h8ffe710_1
libpng=1.6.43=h19919ed_0
libprotobuf=4.25.3=h503648d_0
libsqlite=3.45.3=hcfcfb64_0
libssh2=1.11.0=h7dfc565_0
libtiff=4.6.0=hddb2be6_3
libuv=1.48.0=hcfcfb64_0
libwebp-base=1.4.0=hcfcfb64_0
libxcb=1.15=hcd874cb_0
libxml2=2.12.6=hc3477c8_2
libzlib=1.2.13=hcfcfb64_5
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.5=py38h91455d4_0
matplotlib-base=3.7.3=py38h2724991_0
mkl=2021.4.0=pypi_0
mkl-devel=2021.4.0=h57928b3_730
mkl-include=2021.4.0=h0e2418a_729
mpmath=1.3.0=pyhd8ed1ab_0
msgpack=1.0.8=pypi_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
networkx=3.1=pyhd8ed1ab_0
numpy=1.24.4=py38h1d91fd2_0
opencv=4.9.0=qt6_py38ha3d2b97_612
openexr=3.2.2=h72640d8_1
openh264=2.4.1=h63175ca_0
openjpeg=2.5.2=h3d672ee_0
openssl=3.2.1=hcfcfb64_1
packaging=24.0=pyhd8ed1ab_0
pandas=2.0.3=py38hf08cf0d_1
patsy=0.5.6=pyhd8ed1ab_0
pcre2=10.43=h17e33f8_0
pillow=10.3.0=py38h894f861_0
pip=23.3.1=py38haa95532_0
pixman=0.43.4=h63175ca_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.1=pyhd8ed1ab_0
pooch=1.8.1=pyhd8ed1ab_0
protobuf=5.27.2=pypi_0
psutil=5.9.8=py38h91455d4_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pugixml=1.14=h63175ca_0
py-cpuinfo=9.0.0=pyhd8ed1ab_0
py-opencv=4.9.0=qt6_py38h8ff50eb_612
pyarrow=17.0.0=pypi_0
pyparsing=3.1.2=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
python=3.8.19=h1aa4202_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.3.0=py3.8_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py38h91455d4_1
qt6-main=6.6.3=h5dee92f_0
ray=2.10.0=pypi_0
referencing=0.35.1=pypi_0
requests=2.31.0=pyhd8ed1ab_0
rpds-py=0.19.0=pypi_0
scipy=1.10.1=py38h1aea9ed_3
seaborn=0.13.2=hd8ed1ab_0
seaborn-base=0.13.2=pyhd8ed1ab_0
setuptools=68.2.2=py38haa95532_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=hfb803bf_1
sqlite=3.41.2=h2bbff1b_0
statsmodels=0.14.1=py38he7056a7_0
svt-av1=2.0.0=h63175ca_0
sympy=1.12=pyh04b8f61_3
tbb=2021.13.0=pypi_0
tensorboardx=2.6.2.2=pypi_0
tk=8.6.13=h5226925_1
torchvision=0.18.0=pypi_0
tqdm=4.66.2=pyhd8ed1ab_0
typing_extensions=4.11.0=pyha770c72_0
ucrt=10.0.22621.0=h57928b3_0
ultralytics=8.2.62=pypi_0
ultralytics-thop=2.0.0=pypi_0
unicodedata2=15.1.0=py38h91455d4_0
urllib3=2.2.1=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
werkzeug=3.0.2=pypi_0
wheel=0.41.2=py38haa95532_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
x264=1!164.3095=h8ffe710_2
x265=3.5=h2d74725_3
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_5
zstd=1.5.5=h12be248_0
